# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Taco gaming",
        "description": "Apis",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://code-connoisseurs-laravel.vercel.app/rest"
        }
    ],
    "paths": {
...
        "/categorias/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Mostrar la información de una categoria",
                "operationId": "aa26f53d1c6df2adcf9cef10d1a9964b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Aderson Felix"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Jara Lazaro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "La categoria no existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "La categoria no existe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/desarrolladores/{id}": {
            "get": {
                "tags": [
                    "Desarrolladores"
                ],
                "summary": "Mostrar la información de un desarrollador",
                "operationId": "b720b8276ca125b7192a21ecc9f80ad9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Desarrollador A"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Descripcion del Desarrollador A"
                                        },
                                        "pais": {
                                            "type": "string",
                                            "example": "País A"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "El desarrollador no existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El desarrollador no existe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}